Add field to use uid for resource names
Currently the resources reconciled by the controller will use the name
of the cluster as the name of the resource. This means that it is not
possible to deploy multiple clusters with the same name to the same
virtual data center.

This adds a field to optionally use the uid of the tenant cluster
instead of the name.