Fix incorrect reason waiting for control plane in machine controller
Currently the machine controller uses the waiting for control plane
provider initialized reason when the owner cluster does not have a ready
control plane. This is not intended for general use but for a specific
case on clusters only.

This updatess the machine controller to use the more generic waiting for
control plane available when machines are waiting for the control plane
to be available.