Update tenant cluster controller to replace id in condition errors
Currently errors from cloud director includes a unique id per request.
This means that when the error messages are used in conditions they will
always be considered a new message and update the last transition time.
This in turn will propagate the message to the owner cluster which in
turn will trigger reconciliation of the tenant cluster in an infinite
loop.

This updates the behaviour to attempt to remove any request ids from
the error messages before using them in any conditions.